requirements.txt
Django==5.1.3
djangorestframework==3.16.1
django-filter==24.3
python-dotenv==1.2.1

# Optional (bonus): async tasks / integration readiness
celery==5.6.2
redis==7.1.0

